revert

+++ com.microsoft.mrtk.graphicstools.unity/Runtime/Shaders/GraphicsToolsStandardInput.hlsl
@@ -240,9 +240,9 @@ CBUFFER_START(UnityPerMaterial)
     
     // #if defined(_ROUND_CORNERS) || defined(_BORDER_LIGHT)
     // #if defined(_INDEPENDENT_CORNERS)
-    half4 _RoundCornersRadius;
-    half _RoundCornerRadius;
-    half _RoundCornerMargin;
+    float4 _RoundCornersRadius;
+    float _RoundCornerRadius;
+    float _RoundCornerMargin;
 
     // #if defined(_BORDER_LIGHT)
     half _BorderWidth;
